Extract workspace delete danger zone into DeleteWorkspace component.
